+++ setup.py
@@ -1,11 +1,12 @@
+#!/usr/bin/env python
 # First, we try to use setuptools. If it's not available locally,
 # we fall back on ez_setup.
 try:
-    from setuptools import setup
+    from setuptools import find_packages, setup
 except ImportError:
     from ez_setup import use_setuptools
     use_setuptools()
-    from setuptools import setup
+    from setuptools import find_packages, setup
 
 long_description = ''
 
@@ -23,12 +24,14 @@ with open('requirements.txt') as requirements_file:
 setup(
     name='synapse-orchestrator',
     description='Synapse-based orchestrator for GA4GH workflows',
-    packages=['synorchestrator'],
     url='https://github.com/Sage-Bionetworks/synapse-orchestrator',
     download_url='https://github.com/Sage-Bionetworks/synapse-orchestrator',
     entry_points={
         'console_scripts': 'orchestrate=synorchestrator.__main__:main'
     },
+    package_dir={'': 'src'},
+    packages=find_packages(where='src', exclude=['*.tests.*']),
+    include_package_data=True,
     long_description=long_description,
     install_requires=install_requires,
     setup_requires=['pytest-runner'],
