+++ src/synorchestrator/util.py
@@ -1,5 +1,4 @@
 import json
-import yaml
 import logging
 import datetime as dt
 
@@ -11,19 +10,6 @@ def heredoc(s, inputs_dict):
     import textwrap
     s = textwrap.dedent(s).format(**inputs_dict)
     return s[1:] if s.startswith('\n') else s
-
-
-def get_yaml(filepath):
-    try:
-        with open(filepath, 'r') as f:
-            return yaml.load(f)
-    except IOError:
-        logger.exception("No file found.  Please create: %s." % filepath)
-
-
-def save_yaml(filepath, app_config):
-    with open(filepath, 'w') as f:
-        yaml.dump(app_config, f, default_flow_style=False)
 
 
 def get_json(filepath):
