--- synorchestrator/config.yaml
@@ -1,47 +1,0 @@
-evals:
-  cwl_md5sum:
-    submission_type: params
-    trs_id: dockstore
-    version_id: develop
-    workflow_attachments:
-    - file:///home/quokka/git/workflow-service/testdata/md5sum.input
-    - file:///home/quokka/git/workflow-service/testdata/dockstore-tool-md5sum.cwl
-    workflow_id: ''
-    workflow_jsonyaml: file:///home/quokka/git/workflow-service/testdata/md5sum.cwl.json
-    workflow_type: CWL
-    workflow_url: /home/quokka/git/workflow-service/testdata/md5sum.cwl
-  wdl_UoM_align:
-    submission_type: params
-    trs_id: dockstore
-    version_id: develop
-    workflow_attachments: []
-    workflow_id: ''
-    workflow_jsonyaml: file:///home/quokka/Desktop/topmed-workflows/aligner/u_of_michigan_aligner/u_of_michigan_aligner.json
-    workflow_type: WDL
-    workflow_url: /home/quokka/Desktop/topmed-workflows/aligner/u_of_michigan_aligner/u_of_michigan_aligner.wdl
-toolregistries:
-  dockstore:
-    auth: ''
-    host: dockstore.org:8443
-    proto: https
-workflowservices:
-  arvados-wes:
-    auth: ''
-    auth_type: ''
-    host: wes.qr1hi.arvadosapi.com
-    proto: http
-  aws-toil-server:
-    auth: ''
-    auth_type: ''
-    host: 54.193.12.111:8080
-    proto: http
-  hca-cromwell:
-    auth: ''
-    auth_type: ''
-    host: g0n2qjnu94.execute-api.us-east-1.amazonaws.com/test
-    proto: http
-  local:
-    auth: ''
-    auth_type: ''
-    host: 0.0.0.0:8080
-    proto: http
