--- synorchestrator/submission_queue.json
@@ -1,1 +1,0 @@
-{}
